#4/11

+++ server/controllers/garden.controller.js
@@ -8,17 +8,25 @@ module.exports = {
 };
 
 function listAllGardens(req, res) {
+    userId = req.decoded._id;
     gardenDao.listAllGardens(cb);
     function cb(err, result) {
         if (err) {
             return res.status(500).send(err);
+        }
+        for (var x in result) {
+            if (result[x].user._id.toString() === userId.toString()) {
+                result.isOwner = true;
+            }
         }
         res.status(200).send(result);
     }
 }
 
 function getGarden(req, res) {
+    var userId = req.decoded._id;
     var gardenId = req.params.gardenId;
+
     if (!gardenId) {
         return res.status(400).send("not found");
     }
@@ -29,6 +37,7 @@ function getGarden(req, res) {
         if (err) {
             return res.status(400).send(err);
         }
+        result.isOwner = userId.toString === result.user.toString() ? true : false;
         res.status(200).send(result);
     }
 }
@@ -59,6 +68,7 @@ function createGarden(req, res) {
 }
 
 function updateGarden(req, res) {
+    var userId = req.decoded._id;
     var gardenId = req.params.gardenId;
     if (!gardenId) {
         return res.status(400).send({
@@ -68,20 +78,25 @@ function updateGarden(req, res) {
 
     }
     var gardenInfo = req.body;
-    if (!gardenId) {
-        return res.status(400).send({
-            errCode: 0,
-            errMsg: "Không tìm thấy!"
-        });
-    }
-    gardenDao.updateGarden(gardenId, gardenInfo, cb);
-    function cb(err, result) {
-        if (err) {
-            return res.status(400).send(err);
+    gardenDao.readGarden({
+        _id: gardenId,
+        user: userId
+    }, function (err, result) {
+        if (err || null === result) {
+            return res.status(400).send({
+                errCode: 1,
+                errMsg: "Bạn không phải là chủ vườn!"
+            });
         }
+        gardenDao.updateGarden(gardenId, gardenInfo, cb);
+        function cb(err, result) {
+            if (err) {
+                return res.status(400).send(err);
+            }
 
-        res.status(200).send(result);
-    }
+            res.status(200).send(result);
+        }
+    });
 }
 
 function deleteGarden(req, res) {
@@ -92,12 +107,23 @@ function deleteGarden(req, res) {
             errMsg: "Không tìm thấy!"
         });
     }
-    gardenDao.deleteGarden(gardenId, cb);
-    function cb(err, result) {
-        if (err) {
-            return res.status(400).send(err);
+    gardenDao.readGarden({
+        _id: gardenId,
+        user: userId
+    }, function (err, result) {
+        if (err || null === result) {
+            return res.status(400).send({
+                errCode: 1,
+                errMsg: "Bạn không phải là chủ vườn!"
+            });
         }
+        gardenDao.deleteGarden(gardenId, cb);
+        function cb(err, result) {
+            if (err) {
+                return res.status(400).send(err);
+            }
 
-        res.status(200).send(result);
-    }
+            res.status(200).send(result);
+        }
+    });
 }
